Bring in an interop dir
The interop directory will serve to hold the testing results encountered.

This may grow large over time so I am attempting to keep it organized from the start.

Bring in a README with some information as well as a template of for reporting/recording testing information.

Fill out the README with a few examples.